Develop T3c: the module GUI leaves imageop.c
The last of T3's three extractions: 51 functions, 1706 lines, out of imageop.c
into imageop_gui.c -- header and expander construction, the enable button and
its accels, focus, expansion, rename/duplicate flows, menus, tooltips, the mask
indicator, gui_init/update/cleanup and the bauhaus bridge, joining the
*_from_params factories already there.

Classification differed from the masks cut: both halves are declared in
imageop.h, so declaring-header closure separates nothing. Functions were seeded
by toolkit content (gtk_/Gdk/bauhaus/dtgtk/accels in the body), the closure
pulled in statics referenced only from seeded code, and five orchestrators the
seed regex cannot see -- they only CALL GUI functions, never touch GTK
themselves -- were promoted from the survey's list (gui_duplicate, update_header,
the visible-module walkers, the rename idle). The cross-TU static check came
back clean both ways: no core static is needed by the moved code and no moved
static is referenced from core, so nothing needed exposing.

Two functions that look GUI stay deliberately: _iop_ensure_visible is assigned
into common_fields by the module loader, and default_iop_focus is the
token-pasted default for the iop_focus hook -- referenced by ##-concatenation in
the loader macro, invisible to grep, the exact function-pointer trap
doc/exif-split.md records. Both are T4's business, when the loader stops
populating GUI fields.

Three private #defines and five orphaned static forward declarations travelled
with their users -- the last found by Debug's -Werror=unused-function, which is
what it is for.

imageop.c ends at 1650 lines: module lifecycle, params defaults, introspection,
preset migration, masks/blend sync and hashing, list queries, and the pipeline
defaults -- with a handful of survey-documented leak sites (common_fields
population, the accel-probe GUI instance in _init_module_so) left visible for
T4.

Release, Debug (-Werror), nofeatures all build.
